crud-app/backend/requirements.txt
# backend/requirements.txt
flask
psycopg2-binary
werkzeug